Initial Commit
The code is a Python script for a device that measures CO2, temperature, and humidity levels and displays them on a 4-digit display. It also includes a button that can be used to cycle through the displayed values and toggle a window open/close state. The script uses the scd30_i2c library to interface with an SCD30 CO2 sensor, and the grove.gpio and grove.grove_4_digit_display libraries to interface with a Grove button and 4-digit display, respectively.

The code defines global variables for CO2, temperature, and humidity, as well as a global variable for the display option and a list of display options. It then starts two threads: one to continuously update the display based on the selected display option and the sensor measurements, and another to handle button presses and toggle the window open/close state.